+++ java/src/test/java/io/cucumber/gherkin/GherkinDocumentBuilderTest.java
@@ -4,19 +4,22 @@ import io.cucumber.messages.IdGenerator;
 import io.cucumber.messages.types.Comment;
 import io.cucumber.messages.types.FeatureChild;
 import io.cucumber.messages.types.GherkinDocument;
+import io.cucumber.messages.types.Location;
 import io.cucumber.messages.types.Pickle;
 import io.cucumber.messages.types.TableRow;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class GherkinDocumentBuilderTest {
+class GherkinDocumentBuilderTest {
     private final IdGenerator idGenerator = new IncrementingIdGenerator();
 
     @Test
-    public void is_reusable() {
+    void is_reusable() {
         Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
         TokenMatcher matcher = new TokenMatcher();
 
@@ -28,7 +31,7 @@ public class GherkinDocumentBuilderTest {
     }
 
     @Test
-    public void parses_rules() {
+    void parses_rules() {
         Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
         String data = "" +
                 "Feature: Some rules\n" +
@@ -66,7 +69,7 @@ public class GherkinDocumentBuilderTest {
     }
 
     @Test
-    public void parses_just_comments() {
+    void parses_just_comments() {
         Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
         GherkinDocument doc = parser.parse("# Just a comment", "test.feature");
         List<Comment> children = doc.getComments();
@@ -74,7 +77,7 @@ public class GherkinDocumentBuilderTest {
     }
 
     @Test
-    public void sets_empty_table_cells() {
+    void sets_empty_table_cells() {
         Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
         GherkinDocument doc = parser.parse("" +
                 "Feature:\n" +
@@ -90,7 +93,7 @@ public class GherkinDocumentBuilderTest {
     }
 
     @Test
-    public void table_cells_with_different_size_throws_exception() {
+    void table_cells_with_different_size_throws_exception() {
         // Given
         Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
 
@@ -108,7 +111,7 @@ public class GherkinDocumentBuilderTest {
         assertTrue(compositeParserException.getMessage().contains("inconsistent cell count within the table"));
         Location location = compositeParserException.errors.get(0).location;
         assertEquals(5, location.getLine());
-        assertEquals(7, location.getColumn());
+        assertEquals(7, location.getColumn().get());
     }
 
 }