+++ java/src/test/java/io/cucumber/gherkin/GherkinLineTest.java
@@ -1,5 +1,6 @@
 package io.cucumber.gherkin;
 
+import io.cucumber.messages.types.Location;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
@@ -12,12 +13,12 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class GherkinLineTest {
+class GherkinLineTest {
 
-    private final int line = 12;
+    final Location line = Locations.atLine(12);
 
     @Test
-    public void allows_any_non_space_characters_in_a_tag() {
+    void allows_any_non_space_characters_in_a_tag() {
         GherkinLine gherkinLine = new GherkinLine("   @foo:bar  @zap\uD83E\uDD52yo", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -29,7 +30,7 @@ public class GherkinLineTest {
 
 
     @Test
-    public void finds_tags() {
+    void finds_tags() {
         GherkinLine gherkinLine = new GherkinLine("@this @is @a @tag", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -42,19 +43,19 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void throws_on_tags_with_spaces() {
+    void throws_on_tags_with_spaces() {
         GherkinLine gherkinLine = new GherkinLine("@this @is @a space separated @tag", line);
         assertThrows(ParserException.class, gherkinLine::getTags);
     }
 
     @Test
-    public void throws_on_tags_with_leading_spaces() {
+    void throws_on_tags_with_leading_spaces() {
         GherkinLine gherkinLine = new GherkinLine("@ leadingSpace", line);
         assertThrows(ParserException.class, gherkinLine::getTags);
     }
 
     @Test
-    public void ignores_empty_tag() {
+    void ignores_empty_tag() {
         GherkinLine gherkinLine = new GherkinLine("@", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -62,7 +63,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void ignores_empty_tags() {
+    void ignores_empty_tags() {
         GherkinLine gherkinLine = new GherkinLine("@@", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -70,7 +71,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_tags__trim_whitespace() {
+    void finds_tags__trim_whitespace() {
         GherkinLine gherkinLine = new GherkinLine("    @this @is  @a @tag  ", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -83,7 +84,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_tags__comment_inside_tag() {
+    void finds_tags__comment_inside_tag() {
         GherkinLine gherkinLine = new GherkinLine("@this @is #acomment  ", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -94,7 +95,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_tags__commented_before_tag() {
+    void finds_tags__commented_before_tag() {
         GherkinLine gherkinLine = new GherkinLine("@this @is #@a commented tag", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -105,7 +106,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_tags__commented_multiple_tags() {
+    void finds_tags__commented_multiple_tags() {
         GherkinLine gherkinLine = new GherkinLine("@this @is #@a @commented @sequence of tags", line);
         List<GherkinLineSpan> gherkinLineSpans = gherkinLine.getTags();
 
@@ -116,7 +117,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_table_cells() {
+    void finds_table_cells() {
         // The cells below has the following whitespace characters on each side:
         // - U+00A0 (non-breaking space)
         // - U+0020 (space)
@@ -143,7 +144,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void finds_escaped_table_cells() {
+    void finds_escaped_table_cells() {
         GherkinLine gherkinLine = new GherkinLine("      | \\|æ\\\\n     | \\o\\no\\  | \\\\\\|a\\\\\\\\n | ø\\\\\\nø\\\\|", line);
 
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
@@ -151,7 +152,7 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void preserve_escaped_new_lines_at_start_and_end() {
+    void preserve_escaped_new_lines_at_start_and_end() {
         GherkinLine gherkinLine = new GherkinLine("      |  \nraindrops--\nher last kiss\ngoodbye.\n  |", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("" +
@@ -163,42 +164,42 @@ public class GherkinLineTest {
     }
 
     @Test
-    public void escapes_backslash() {
+    void escapes_backslash() {
         GherkinLine gherkinLine = new GherkinLine("|\\\\o\\no\\||", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("\\o\no|"), texts);
     }
 
     @Test
-    public void throws_on_illegal_escapes_backslash() {
+    void throws_on_illegal_escapes_backslash() {
         GherkinLine gherkinLine = new GherkinLine("|\\o\\no\\||", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("\\o\no|"), texts);
     }
 
     @Test
-    public void correctly_trims_white_spaces_before_cell_content() {
+    void correctly_trims_white_spaces_before_cell_content() {
         GherkinLine gherkinLine = new GherkinLine("|   \t spaces before|", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces before"), texts);
     }
 
     @Test
-    public void correctly_trims_white_spaces_after_cell_content() {
+    void correctly_trims_white_spaces_after_cell_content() {
         GherkinLine gherkinLine = new GherkinLine("|spaces after   |", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces after"), texts);
     }
 
     @Test
-    public void correctly_trims_white_spaces_around_cell_content() {
+    void correctly_trims_white_spaces_around_cell_content() {
         GherkinLine gherkinLine = new GherkinLine("|   \t spaces everywhere   \t|", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces everywhere"), texts);
     }
 
     @Test
-    public void does_not_drop_white_spaces_inside_a_cell() {
+    void does_not_drop_white_spaces_inside_a_cell() {
         GherkinLine gherkinLine = new GherkinLine("| foo()\n  bar\nbaz |", line);
         List<String> texts = gherkinLine.getTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("foo()\n  bar\nbaz"), texts);